splitting off tests into separate file
