+++ include/errors.h
@@ -7,6 +7,17 @@
 #include "macros.h"
 #include <CL/cl.h>
 
+#include <stdio.h>
+
+// QBT -> Quibble Text
+#define QBT_RED     "\x1b[31m"
+#define QBT_GREEN   "\x1b[32m"
+#define QBT_YELLOW  "\x1b[33m"
+#define QBT_BLUE    "\x1b[34m"
+#define QBT_MAGENTA "\x1b[35m"
+#define QBT_CYAN    "\x1b[36m"
+#define QBT_RESET   "\x1b[0m"
+
 void cl_check(cl_int error);
 
 const char *get_cl_error_string(cl_int error);
