+++ include/quibble_stanzas.c
@@ -1,0 +1,4 @@
+#ifndef QUIBBLE_STANZAS
+#define QUIBBLE_STANZAS
+
+#endif
