+++ include/tests.h
@@ -1,0 +1,8 @@
+#ifndef QUIBBLE_TESTS
+#define QUIBBLE_TESTS
+
+#include "quibble_verses.h"  
+
+void quibble_verse_tests();
+
+#endif
